Separates specs into json and yaml only
